% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disp_est_null_family.R
\name{Disp_est_null_family}
\alias{Disp_est_null_family}
\title{Estimate Dispersion Parameters Utilizing Null Families in A Gene Cluster.}
\usage{
Disp_est_null_family(counts, group = NULL)
}
\arguments{
\item{counts}{a count matrix with genes in rows, and three*n_families columns, female, male and hybrid which passed Pois_3mean_LRT() and can be treated as null families.}

\item{group}{a vector indicating the group information of three*n_families columns, usually one null family is one group.}
}
\value{
a vector containing gene-wise dispersion estimation.
}
\description{
This function uses edgeR to estimate the tagwise dispersion parameter for multiple genes (from a gene cluster) with multiple family (usually the null families detected by Pois_3mean_LRT()).
}
\examples{
library(TwoStageLRT)
counts <- matrix(rpois(150, 10), nrow = 10, ncol = 15)
group <- rep(1:5, each = 3)
disp_est <- Disp_est_null_family(counts = counts, group = group)
summary(disp_est)
}
