% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hetero_NB_LRT.R
\name{Hetero_NB_LRT}
\alias{Hetero_NB_LRT}
\title{Negative Binomial Likelihood Ratio Test with Known Dispersion for Detection of Mid-parent Heterosis}
\usage{
Hetero_NB_LRT(y_g1, y_g2, y_g3, C_1 = 1, C_2 = 2, C_3 = 1, dispersion = 1)
}
\arguments{
\item{y_g1}{raw count for female.}

\item{y_g2}{raw count for male.}

\item{y_g3}{raw count for hybrid.}

\item{C_1}{normalization factor for female sample.}

\item{C_2}{normalization factor for male sample.}

\item{C_3}{normalization factor for hybrid sample.}

\item{dispersion}{value of dispersion parameter.}
}
\value{
p-value.
}
\description{
This function conducts negative binomial LRT with known dispersion for detection of gene expression mid-parent heterosis (MPH) in a single family with two parental genotypes and one hybrid genotype and for a single gene with one sample per genotype.
Denote the mean for gene g and female parent as mu_g1, male parent as mu_g2 and hybrid as mu_g3, the null hypothesis is Hg0: mu_g1 + mu_g2 = 2 * mu_g3.
Refer the paper for details of this test, since there is no closed form for the maximum likelihood estimate under null hypothesis, numerical algorithms are used with the global MLE as the starting points. Algorithms will fail if any of the raw counts is 0.
}
\examples{
library(TwoStageLRT)
Hetero_NB_LRT(y_g1 = 10, y_g2 = 15, y_g3 = 20, C_1 = 1, C_2 = 1.2, C_3 = 0.8, dispersion = 0.1)
}
