% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_func.R
\name{inverse_func}
\alias{inverse_func}
\title{Simulation dispersion parameters as a function of mid-parent mean}
\usage{
inverse_func(mp)
}
\arguments{
\item{mp}{mid-parent mean.}
}
\value{
simulated dispersion parameter value.
}
\description{
disp = (mp)^(-1/6) - 0.3 + rnorm(1, 0, sd = 0.002)
}
\examples{
library(TwoStageLRT)
inverse_func(100)
}
