% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Simulated unreplicated multi-familiy RNA-seq expression data.}
\format{
A list containing elements female_mat, male_mat, hybrid_mat, disp_mat and norm_factors.
}
\usage{
sim_data
}
\description{
A dataset including gene exprssion for 100 genes and 150 families. Each family has three genotypes, female as columns in \code{female_mat}, male as columns in \code{male_mat}, hybrid as columns in \code{hybrid_mat}, true genewise dispersion parameters in \code{disp_mat}, true normalization factors in \code{norm_factors}.
}
\keyword{sim_data}
