% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoStageLRT.R
\name{twoStageLRT}
\alias{twoStageLRT}
\title{Two Stage Likelihood Ratio Test for Detection of Gene Expression Mid-parent Heterosis}
\usage{
twoStageLRT(
  hybrid_mat,
  female_mat,
  male_mat,
  disp_mat = NULL,
  normalization = FALSE,
  norm_factors = NULL,
  n_fam_thres = 20,
  gene_groups_no = 3
)
}
\arguments{
\item{hybrid_mat}{hybrids expression count with genes in rows, samples in columns.}

\item{female_mat}{female expression count with genes in rows, samples in columns.}

\item{male_mat}{male expression count with genes in rows, samples in columns.}

\item{disp_mat}{if "Pois", use Pois LRT and only output Pois results; if NULL, use dispersion estimation  output will have both NBLRT and PoisLRT results; else, use provided disp_mat for NB test, output will have both NBLRT and PoisLRT results.}

\item{normalization}{logical, whether or not need normalization for the raw data matrices, if FALSE, treat the normalization factors all 1.}

\item{norm_factors}{if normalization = TRUE and norm_factors = NULL, use DESeq normalization factors; if normalization = TRUE and norm_factors is a list contains three vectors female, male and hybrid normalization factors, use the provided norm_factors in tests.}

\item{n_fam_thres}{genes with at least n_fam_thres null families will be used for dispersion estimation, otherwise, will use loess prediction to get dispersion, default is 20.}

\item{gene_groups_no}{the number of clusters for genes, default is 3.}
}
\value{
if disp_mat = "Pois", return Poisson LRT for heterosis pvalue and adjusted pvalue from BH adjustment; if disp_mat = NULL or provided, return both NB and Poisson LRT for heterosis pvalue and qvalue and genewise dispersion, and the estimation of mu_gi, which is the female/male/hybrid_mat_normed.
}
\description{
This function apply two stage likelihood ratio test (2sLRT) to multi-family RNA-seq data with formats: three count matrices for famale, male and hybrid with genes in rows, samples in columns.
}
\examples{
library(TwoStageLRT)
data(sim_data)
res <- twoStageLRT(female_mat = sim_data$female_mat, male_mat = sim_data$male_mat,
  hybrid_mat = sim_data$hybrid_mat, disp_mat = NULL,
  normalization = TRUE, norm_factors = NULL,
  n_fam_thres = 10, gene_groups_no = 2)
summary(res$pval_NB)
summary(res$pval_Pois)
summary(res$DISP)
}
